produces no move events on conditional writes with false predicates
We represent conditional writes as `r := if c then x else r` so that
if `c` is false the contents of `r` doesn't change. However, it was
producing a write event that was triggering a false event
mismatching. To prevent this, we check on each update that the new
value has a different id and do not emit any events if they are the
same.

Also, improves some pretty printing functions to make the errors
output more readable and easier to use with bap mc.