adds support for mode events (#24)
* adds support for mode events

Mode events from bap-frames traces, indicating for example whether an
instruction is executed in thumb mode or not, are now taken into account
for disassembly and lifting. This enables chcecking traces from qemu
running arm code with interworking.

As a consequence of possibly requiring different disassemblers per
frame and respecting the mode, the code has been refactored to locally
use the KB to derive target and encoding for `Dis.lookup` instead of
running everything in `Dis.with_disasm`.

* fixes desugaring of var aliases

We need to promise a unit for newly created labels during lifting,
otherwise the `Desugar` pass will not be able to infer the target,
making it fail. This would be reflected in aarch64 BIL having the
X0/SP/... aliases instead of the R0/R31/... we need.